+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Breaking changes:
     compatible with all requests from ruma-client-api and ruma-appservice-api.
   - `HttpRequest::RequestBuilder` has an extra `AsRef<[u8]>` bound.
   - Bump MSRV to 1.88
-- `Client::send_request_as()` now takes `AppserviceUserIdentity` instead of `&UserId`.
+  - `Client::send_request_as()` and `HttpClientExt::send_matrix_request_as()` now take `AppserviceUserIdentity`
+    instead of `&UserId`.
   
 Improvements: